create_plugin_registries now defaults to pickle instead of yaml
