+++ geoips/geoips_utils.py
@@ -15,7 +15,7 @@
 import os
 from copy import deepcopy
 import sys
-import yaml
+# import yaml
 import logging
 from importlib import metadata, resources
 
@@ -73,11 +73,12 @@ def load_all_yaml_plugins():
     in ``.yaml``. Read each plugin file
     """
     # Load all entry points for plugin packages
+    import pickle
 
     plugin_packages = get_entry_point_group("geoips.plugin_packages")
     yaml_plugins = {}
     for pkg in plugin_packages:
-        pkg_plug_path = str(resources.files(pkg.value) / "registered_plugins.yaml")
+        pkg_plug_path = str(resources.files(pkg.value) / "registered_plugins")
         if not os.path.exists(pkg_plug_path):
             raise PluginRegistryError(
                 f"Plugin registry {pkg_plug_path} did not exist, "
@@ -85,7 +86,7 @@ def load_all_yaml_plugins():
             )
         # This will include all plugins, including schemas, yaml_based,
         # and module_based plugins.
-        registered_plugins = yaml.safe_load(open(pkg_plug_path, "r"))
+        registered_plugins = pickle.load(open(pkg_plug_path, "rb"))
         # Only pull the "yaml_based" plugins here.
         try:
             for interface in registered_plugins["yaml_based"]:
